Set pickup/drop_off _type on in-seat transfers
If a transfer is in_seat, passengers should be able to embark at the
last stop of the from_trip of the continuation, and disembark at the
first stop of the to_trip of the continuation. Update the stop_times to
reflect this.

Update a simple test case to ensure that this transformation is applied.